+++ docs/sphinx/conf.py
@@ -13,6 +13,8 @@
 # serve to show the default.
 
 import sphinx_bootstrap_theme
+import sys
+import os
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
@@ -21,6 +23,8 @@ import sphinx_bootstrap_theme
 from sdss_access import __version__
 from pkg_resources import parse_version
 
+# add a local path to the Sphinx search path
+sys.path.insert(0, os.path.abspath('../../python/sdss_access/misc/'))
 
 # -- General configuration ------------------------------------------------
 
@@ -33,7 +37,7 @@ from pkg_resources import parse_version
 # ones.
 extensions = ['sphinx.ext.autodoc', 'sphinx.ext.napoleon', 'sphinx.ext.autosummary',
               'sphinx.ext.todo', 'sphinx.ext.viewcode', 'sphinx.ext.mathjax',
-              'sphinx.ext.intersphinx']
+              'sphinx.ext.intersphinx', 'docupaths']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']