+++ CHANGELOG.rst
@@ -9,6 +9,11 @@ This document records the main changes to the sdss_access code.
 0.2.4 (unreleased)
 ------------------
 
+Added
+-----
+* Method to lookup the keyword arguments needed for a given path name
+* Sphinx plugin to auto document the sdss_access path definitions
+
 .. _changelog-0.2.3:
 
 0.2.3 (2017-12-02)